+++ src/components/common/about/section/solution/SolutionSection.tsx
@@ -1,3 +1,5 @@
+import SolutionBox from './SolutionBox';
+
 const CONTENT = [
   {
     imgName: 'lets-solution.svg',
@@ -38,18 +40,8 @@ const SolutionSection = () => {
         </p>
       </div>
       <div className="flex max-w-[63.75rem] flex-col gap-20 md:flex-row md:gap-[3.56rem] xl:gap-[7.5rem]">
-        {CONTENT.map(({ title, description, imgName, imgWidth }) => (
-          <div key={title} className="flex items-center gap-6 md:flex-col">
-            <img className={`h-full ${imgWidth}`} src={`/icons/${imgName}`} />
-            <div className="flex flex-col gap-2">
-              <h2 className="text-1-bold xl:text-1.25-bold text-primary md:text-center">
-                {title}
-              </h2>
-              <p className="text-0.875 xl:text-1.125 text-justify text-neutral-40 md:text-center">
-                {description}
-              </p>
-            </div>
-          </div>
+        {CONTENT.map((item) => (
+          <SolutionBox key={item.title} {...item} />
         ))}
       </div>
     </section>
