+++ src/pages/common/about/About.tsx
@@ -1,4 +1,5 @@
 import AboutHeader from '../../../components/common/about/header/AboutHeader';
+import ResultSection from '../../../components/common/about/ResultSection';
 import CommunitySection from '../../../components/common/about/section/CommunitySection';
 import ContactSection from '../../../components/common/about/section/ContactSection';
 import EndSection from '../../../components/common/about/section/EndSection';
@@ -6,9 +7,8 @@ import IntroSection from '../../../components/common/about/section/introduction/
 import PartnerSection from '../../../components/common/about/section/PartnerSection';
 import ProblemSection from '../../../components/common/about/section/problem/ProblemSection';
 import ProgramMenuSection from '../../../components/common/about/section/program/ProgramMenuSection';
-import ResultSection from '../../../components/common/about/section/result/ResultSection';
 import ReviewSection from '../../../components/common/about/section/ReviewSection';
-import SolutionSection from '../../../components/common/about/section/SolutionSection';
+import SolutionSection from '../../../components/common/about/section/solution/SolutionSection';
 
 const About = () => {
   return (
